# src/main/java/za/co/java/engineer/raymond/shape/api/model/Shape.java
package za.co.java.engineer.raymond.shape.api.model;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.UniqueConstraint;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
public class Shape {
    @Id
    private String name;
    private String type;
    private String description;
}